Create separate sidebar button component
