+++ app/src/components/intake-profile/ProfileSidebar.tsx
@@ -1,7 +1,6 @@
-import {Box, Stack, Typography, useTheme} from '@mui/material';
-import {Link} from 'react-router-dom';
-import useStatusStyling from '../../views/IntakeProfile/hooks/useStatusStyling';
+import {Stack, Typography} from '@mui/material';
 import {Dispatch, SetStateAction} from 'react';
+import {SidebarButton} from './SidebarButton';
 
 type FieldGroup = {
   id: string;
@@ -22,8 +21,6 @@ export const ProfileSidebar = ({
   groupId,
 }: ProfileSidebarProps) => {
   const totalTask = fieldGroups.length - 1;
-  const statusStyling = useStatusStyling();
-  const theme = useTheme();
 
   const handleItemClick = (id: string) => {
     setSelectedItem(id);
@@ -48,43 +45,18 @@ export const ProfileSidebar = ({
         </Typography>
       </Stack>
       {fieldGroups.map(({id, title}) => {
-        const status = 'complete'; // Change status here to see different styles
+        const status = 'incomplete'; // Change status here to see different styles
         // complete | partial | incomplete | locked
         const fieldTitle = title || '...';
         return (
-          <Box
+          <SidebarButton
             key={id}
-            to={`group/${id}`}
-            component={Link}
-            sx={{
-              borderRadius: 2,
-              backgroundColor: statusStyling[status].color,
-              minHeight: 56,
-              display: 'flex',
-              justifyContent: 'space-between',
-              alignItems: 'center',
-              paddingInline: 3,
-              borderColor:
-                groupId === id ? '#DADADA' : statusStyling[status].borderColor,
-              borderWidth: 2,
-              borderStyle: 'solid',
-              boxShadow:
-                groupId === id ? '0px 4px 4px rgba(0, 0, 0, .25)' : 'none',
-              textDecoration: 'none',
-            }}
-            onClick={() => handleItemClick(id)}
-          >
-            <Typography
-              sx={{
-                fontSize: 16,
-                fontWeight: 'medium',
-                color: theme.palette.text.primary,
-              }}
-            >
-              {fieldTitle}
-            </Typography>
-            {statusStyling[status].icon}
-          </Box>
+            fieldTitle={fieldTitle}
+            groupId={groupId}
+            handleClick={handleItemClick}
+            id={id}
+            status={status}
+          />
         );
       })}
     </>
